add index.js index.css files
