+++ modules/getbooks.js
@@ -1,0 +1,1 @@
+export const getBooks = () => JSON.parse(localStorage.getItem('bookData')) || [];
