+++ modules/savebooks.js
@@ -1,0 +1,1 @@
+export const saveBooks = arrBooks => localStorage.setItem('bookData', JSON.stringify(arrBooks));
